More attribute refactoring; changing how attribute values work for backward compatibility
Former-commit-id: 7c7343fe6d53a2fe042682f533b029ff41aae71f
Former-commit-id: 276fbbb08ff5eae8433d5a76239555adfa3e413f

+++ concrete/src/Attribute/Category/AbstractCategory.php
@@ -236,13 +236,13 @@ abstract class AbstractCategory implements CategoryInterface, StandardSearchInde
          * @var Value
          */
         $value = $attribute->getValueObject();
-        $this->entityManager->remove($attribute);
-
-        $this->entityManager->flush();
-
-        $this->entityManager->refresh($value);
-        if (count($value->getAttributeValues()) < 1) {
-            $this->entityManager->remove($value);
+        if (is_object($value)) {
+            $this->entityManager->remove($attribute);
+            $this->entityManager->flush();
+            $this->entityManager->refresh($value);
+            if (count($value->getAttributeValues()) < 1) {
+                $this->entityManager->remove($value);
+            }
         }
     }
